+++ androidApp/src/androidMain/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">My application</string>
+    <string name="app_name">Compose Multiplatform Exercise</string>
 </resources>
