+++ iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,3 @@
 TEAM_ID=
-BUNDLE_ID=com.myapplication.MyApplication
-APP_NAME=My application
+BUNDLE_ID=net.wwwhackcom.ComposeMultiplatform
+APP_NAME=Compose Multiplatform application
